Show incident assigned user in list view
